15Task. FitnessPal/ resize widget to 5x1
