+++ src/main/java/koval/mobile/myfitnesspal/gui/MyAbstractPage.java
@@ -1,60 +1,15 @@
 package koval.mobile.myfitnesspal.gui;
 
-import com.qaprosoft.carina.core.foundation.webdriver.decorator.ExtendedWebElement;
-import com.qaprosoft.carina.core.gui.AbstractPage;
-import org.openqa.selenium.*;
-import org.openqa.selenium.support.FindBy;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import com.zebrunner.carina.utils.factory.DeviceType;
+import org.openqa.selenium.WebDriver;
 
 
-import java.lang.invoke.MethodHandles;
 
-
-public abstract class MyAbstractPage extends AbstractPage implements IMyInterface {
-
-    @FindBy(xpath = "//*[@content-desc='%s']")
-    private ExtendedWebElement itemByContent;
-
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+@DeviceType(pageType = DeviceType.Type.ANDROID_PHONE, parentClass = MyAbstractPageBase.class)
+public class MyAbstractPage extends MyAbstractPageBase {
 
     public MyAbstractPage(WebDriver driver) {
         super(driver);
     }
-
-
-    public boolean isPageOpened(ExtendedWebElement title, String textTitle) {
-        return title.isElementPresent(TIMEOUT_FIVE) && title.getText().equals(textTitle);
-    }
-
-    public void hideKeyboard() {
-        pressKey(BACK);
-    }
-
-    public void pressKey(int keyCodeValue) {
-        getDevice(getDriver()).pressKey(keyCodeValue);
-    }
-
-    public void closeTimestampsPopUpIfPresent() {
-        itemByContent.format(NO_THANKS_ANSWER.toUpperCase()).clickIfPresent(TIMEOUT_FIVE);
-    }
-
-
-    public int getCenterX(ExtendedWebElement extendedWebElement) {
-
-        int leftX = extendedWebElement.getLocation().getX();
-        int rightX = extendedWebElement.getSize().getWidth();
-
-        return (leftX + (leftX + rightX)) / 2;
-    }
-
-    public int getCenterY(ExtendedWebElement extendedWebElement) {
-
-        int upperY = extendedWebElement.getLocation().getY();
-        int lowerY = extendedWebElement.getSize().getHeight();
-
-        return (upperY + (upperY + lowerY)) / 2;
-    }
-
 }
